pyproject.toml
[project]
name = "auto-snap-mcp"
version = "0.1.0"
description = "MCP server for automated screenshot capture and PDF conversion"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "img2pdf>=0.5.0",
    "pyscreenshot>=3.1.0"
]
